refactor: splits new product functionality into separate component
